Add async versions of `caching` and `cachingWithTTL`
This is a breaking API change because the existing methods have been
renamed to `cachingSync` and `cachingSyncWithTTL`.

The opinionated stance of ScalaCache is that you should use the async
API and everything should be a Future. We also provide synchronous
counterparts as a convenience, but the methods get slightly uglier
names.