WIP experimenting with giving Cache a higher-kinded type param
Goals:

* Improve performance of in-memory caches by removing unnecessary Future overhead
* Get rid of the redundant sync API and replace with modes

The idea is that you'll be able to choose whether you want to be sync or
async by importing a mode, inspired by Rapture. Yet to try this: not
sure whether it will work as I hope.

TODO:

* Experiment with modes
* Make RedisCache and MemcachedCache compile
* Update the memoize macros to match the new API
* Update docs and write migration guide

If all of this goes to plan, the only things standing between us and a
1.0 release are:

* Circe codecs
* A docs site