Returns combined Future[From] results on cache
Combine cache putWithKey result with the original
body result futures with flatMap, allowing the caller to be aware
of when the data gets actually cached and assure consistency of
expected behaviour.

Without this code change, some paradox situations were actually
possible where data was supposed to be cached but it wasn't.

Example of a failing test:

```
    val results = for {
      res1 <- cache.caching("key")(Future { "initial value" })
      res2 <- cache.caching("key")(Future { "updated value" })
    } yield (Seq(res1, res2))

    results map {
      _ should contain only ("initial value")
    }
```